# backend/docs/swagger.yaml
basePath: /
definitions:
  backend_internal_model.Folder:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_folder_id:
        description: nil for root folders
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  backend_internal_model.Item:
    properties:
      id:
        type: string
      name:
        type: string
      type:
        description: '"note" or "folder"'
        type: string
    type: object
  backend_internal_model.Note:
    properties:
      content:
        type: string
      created_at:
        type: string
      folder_id:
        description: nil for notes outside any folder
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  backend_internal_pkg.AuthResponse:
    properties:
      access_token:
        type: string
      token:
        type: string
    type: object
  backend_internal_pkg.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  backend_internal_pkg.RefreshTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  backend_internal_pkg.SignupResponse:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_handlers.CreateFolderRequest:
    properties:
      name:
        type: string
      parent_folder_id:
        type: string
      user_id:
        type: string
    required:
    - name
    - user_id
    type: object
  internal_handlers.CreateNoteRequest:
    properties:
      content:
        type: string
      folder_id:
        description: nil to create the note outside any folder
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  internal_handlers.DeleteFoldersRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  internal_handlers.DeleteNotesRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  internal_handlers.FolderResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/backend_internal_model.Item'
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_folder_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_handlers.UpdateFolderItem:
    properties:
      id:
        type: string
      name:
        type: string
      parent_folder_id:
        type: string
    required:
    - id
    type: object
  internal_handlers.UpdateFoldersRequest:
    properties:
      folders:
        items:
          $ref: '#/definitions/internal_handlers.UpdateFolderItem'
        minItems: 1
        type: array
    required:
    - folders
    type: object
  internal_handlers.UpdateNoteItem:
    properties:
      content:
        type: string
      id:
        type: string
      title:
        type: string
    required:
    - id
    type: object
  internal_handlers.UpdateNotesRequest:
    properties:
      notes:
        items:
          $ref: '#/definitions/internal_handlers.UpdateNoteItem'
        minItems: 1
        type: array
    required:
    - notes
    type: object
info:
  contact: {}
  description: API for managing users, notes, and folders.
  title: Note Taking App API
  version: "1.0"
paths:
  /api/folders:
    delete:
      consumes:
      - application/json
      description: Deletes one or more folders owned by the authenticated user in
        a single batch
      parameters:
      - description: Folder IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DeleteFoldersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete multiple folders
      tags:
      - folders
    get:
      description: Returns all folders owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_model.Folder'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List folders
      tags:
      - folders
    post:
      consumes:
      - application/json
      description: Creates a new folder, optionally nested inside a parent folder
      parameters:
      - description: Folder to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_model.Folder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a folder
      tags:
      - folders
    put:
      consumes:
      - application/json
      description: Updates the name and/or parent of one or more folders in a single
        batch
      parameters:
      - description: Folders to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateFoldersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_model.Folder'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update multiple folders
      tags:
      - folders
  /api/folders/{id}:
    get:
      description: Returns a folder along with its direct child folders and notes
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.FolderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a folder
      tags:
      - folders
  /api/notes:
    delete:
      consumes:
      - application/json
      description: Deletes one or more notes owned by the authenticated user in a
        single batch
      parameters:
      - description: Note IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DeleteNotesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete multiple notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Creates a new note, optionally inside a folder
      parameters:
      - description: Note to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_model.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Updates the title and/or content of one or more notes in a single
        batch
      parameters:
      - description: Notes to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateNotesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_model.Note'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update multiple notes
      tags:
      - notes
  /api/notes/{id}:
    get:
      description: Returns a single note owned by the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_model.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access/refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_pkg.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_pkg.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns access/refresh tokens
      parameters:
      - description: Signup details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_pkg.SignupResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign up
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"